requirements.txt
# Production dependencies
click>=8.0.0
typing-extensions>=4.0.0
Pillow

# Development dependencies  
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0